Added Aspire.Hosting.Azure.Functions
