+++ src/Aspire.Hosting.Azure.Functions/PublicAPI.Shipped.txt
@@ -1,0 +1,1 @@
+#nullable enable
