+++ src/Aspire.Hosting.Azure.Functions/PublicAPI.Unshipped.txt
@@ -1,0 +1,8 @@
+#nullable enable
+Aspire.Hosting.Azure.AzureFunctionsProjectResource
+Aspire.Hosting.Azure.AzureFunctionsProjectResource.AzureFunctionsProjectResource(string! name, string! executable, string! workingDirectory) -> void
+Aspire.Hosting.Azure.AzureFunctionsProjectResourceExtensions
+static Aspire.Hosting.Azure.AzureFunctionsProjectResourceExtensions.AddAzureFunctionsProject<TProject>(this Aspire.Hosting.IDistributedApplicationBuilder! builder, string! name) -> Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureFunctionsProjectResource!>!
+static Aspire.Hosting.Azure.AzureFunctionsProjectResourceExtensions.WithReference(this Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureFunctionsProjectResource!>! builder, Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureBlobStorageResource!>! source, string? connectionName = null) -> Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureFunctionsProjectResource!>!
+static Aspire.Hosting.Azure.AzureFunctionsProjectResourceExtensions.WithReference(this Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureFunctionsProjectResource!>! builder, Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureEventHubsResource!>! source, string? connectionName = null) -> Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureFunctionsProjectResource!>!
+static Aspire.Hosting.Azure.AzureFunctionsProjectResourceExtensions.WithReference(this Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureFunctionsProjectResource!>! builder, Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureQueueStorageResource!>! source, string? connectionName = null) -> Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureFunctionsProjectResource!>!
