+++ src/Aspire.Hosting.Azure.Functions/README.md
@@ -1,0 +1,25 @@
+# Aspire.Hosting.Azure.Functions library
+
+Provides methods to the .NET Aspire hosting model for Azure functions.
+
+## Getting started
+
+### Install the package
+
+In your AppHost project, install the .NET Aspire Azure Functions Hosting library with [NuGet](https://www.nuget.org):
+
+```dotnetcli
+dotnet add package Aspire.Hosting.Azure.Functions
+```
+
+## Usage example
+
+Then, in the _Program.cs_ file of `AppHost`, add a resource based on a Bicep template:
+
+```csharp
+
+```
+
+## Feedback & contributing
+
+https://github.com/dotnet/aspire
